backend/requirements.txt
Flask==2.0.3
Flask-Cors==3.0
mysql-connector-python
Pillow